improve networking, use view model and delegate

+++ Weather Practice/Weather Practice/WeatherModel.swift
@@ -17,6 +17,6 @@ struct CurrentWeatherResponse: Decodable {
     }
     
     struct Main: Decodable {
-        let temp: Decimal
+        let temp: Double
     }
 }
